# spring-api/openapi/hikingbuddy/src/main/resources/hikingbuddy.yml
openapi: 3.0.1
info:
  title: GroupProjectApi
  version: v1
paths:
...
  /api/Friends/accepted/users/{email}:
    get:
      tags:
        - Friends
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
            text/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
...
  /api/Friends/incoming/users/{email}:
    get:
      tags:
        - Friends
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
            text/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
...
  /api/Friends/outgoing/users/{email}:
    get:
      tags:
        - Friends
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
            text/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
  /api/Friends/{email1}/{email2}:
    get:
      tags:
        - Friends
      parameters:
        - name: email1
          in: path
          required: true
          schema:
            type: string
            nullable: true
        - name: email2
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/FriendDto"
            application/json:
              schema:
                $ref: "#/components/schemas/FriendDto"
            text/json:
              schema:
                $ref: "#/components/schemas/FriendDto"
...
  /api/Friends/{email}:
    delete:
      tags:
        - Friends
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/FriendDto"
            application/json:
              schema:
                $ref: "#/components/schemas/FriendDto"
            text/json:
              schema:
                $ref: "#/components/schemas/FriendDto"
...
  /api/UserEvents/users/{eventID}:
    get:
      tags:
        - UserEvents
      parameters:
        - name: eventID
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublicDto"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublicDto"
            text/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublicDto"
...
  /api/Users/{email}:
    get:
      tags:
        - Users
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/UserDto"
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
            text/json:
              schema:
                $ref: "#/components/schemas/UserDto"
...
  /api/Users:
    post:
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
          text/json:
            schema:
              $ref: "#/components/schemas/User"
          application/*+json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/UserDto"
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
            text/json:
              schema:
                $ref: "#/components/schemas/UserDto"
    put:
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDto"
          text/json:
            schema:
              $ref: "#/components/schemas/UserDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UserDto"
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/UserDto"
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
            text/json:
              schema:
                $ref: "#/components/schemas/UserDto"
    delete:
      tags:
        - Users
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/UserDto"
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
            text/json:
              schema:
                $ref: "#/components/schemas/UserDto"
...
components:
  schemas:
...
    UserPublicDto:
      required:
        - dateOfBirth
        - email
        - location
        - name
        - pets
        - skillLevel
      type: object
      properties:
        email:
          maxLength: 30
          minLength: 0
          type: string
          format: email
        name:
          maxLength: 30
          minLength: 0
          type: string
        description:
          type: string
          format: string
          nullable: true
        dateOfBirth:
          type: string
          format: date
        skillLevel:
          maximum: 5
          minimum: 1
          type: integer
          format: int32
        experience:
          type: string
          format: string
          nullable: true
        location:
          maxLength: 6
          minLength: 0
          type: string
          format: postal-code
        pets:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        profilePicture:
          maxLength: 100
          minLength: 0
          type: string
          format: uri
          nullable: true
      additionalProperties: false
    FriendDto:
      required:
        - friendStatus
        - sourceUser
        - targetUser
      type: object
      properties:
        sourceUser:
          type: string
        targetUser:
          type: string
        friendStatus:
          maximum: 1
          minimum: 0
          type: integer
          format: int32
        requestedAt:
          type: string
          format: date-time
      additionalProperties: false
    UserDto:
      required:
        - createdAt
        - dateOfBirth
        - email
        - location
        - name
        - pets
        - phoneNumber
        - skillLevel
        - updatedAt
      type: object
      properties:
        email:
          maxLength: 30
          minLength: 0
          type: string
          format: email
        name:
          maxLength: 30
          minLength: 0
          type: string
        dateOfBirth:
          type: string
          format: date
        skillLevel:
          maximum: 5
          minimum: 1
          type: integer
          format: int32
        location:
          maxLength: 6
          minLength: 0
          type: string
          format: postal-code
        phoneNumber:
          maxLength: 12
          minLength: 0
          type: string
          format: tel
        pets:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        description:
          type: string
          format: string
          nullable: true
        experience:
          type: string
          format: string
          nullable: true
        profilePicture:
          maxLength: 100
          minLength: 0
          type: string
          format: uri
          nullable: true
      additionalProperties: false
...
    User:
      required:
        - dob
        - email
        - fullname
        - location
        - pet
        - phoneNum
        - pwd
        - skillLvl
      type: object
      properties:
        email:
          maxLength: 30
          minLength: 0
          type: string
          format: email
        pwd:
          maxLength: 60
          minLength: 0
          type: string
        fullname:
          maxLength: 30
          minLength: 0
          type: string
        dob:
          type: string
          format: date
        skillLvl:
          maximum: 5
          minimum: 1
          type: integer
          format: int32
        phoneNum:
          maxLength: 11
          minLength: 0
          type: string
          format: tel
        pet:
          type: boolean
        location:
          maxLength: 6
          minLength: 0
          type: string
          format: postal-code
        descript:
          type: string
          format: string
          nullable: true
        experience:
          type: string
          format: string
          nullable: true
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
        profilePic:
          maxLength: 100
          minLength: 0
          type: string
          format: uri
          nullable: true
      additionalProperties: false